label trait feature now part of notable traits features
