+++ compiler/rustc_span/src/symbol.rs
@@ -842,7 +842,6 @@ symbols! {
         doc_cfg,
         doc_cfg_hide,
         doc_keyword,
-        doc_label_trait,
         doc_masked,
         doc_notable_trait,
         doc_primitive,
@@ -1171,7 +1170,6 @@ symbols! {
         kreg0,
         label,
         label_break_value,
-        label_trait,
         lahfsahf_target_feature,
         lang,
         lang_items,
