+++ src/librustdoc/clean/mod.rs
@@ -2820,7 +2820,6 @@ fn add_without_unwanted_attributes<'hir>(
                     first_span: _,
                     aliases,
                     hidden,
-                    label_trait: _,
                     inline,
                     cfg,
                     auto_cfg: _,
