+++ src/librustdoc/clean/types.rs
@@ -1365,9 +1365,6 @@ impl Trait {
     pub(crate) fn is_auto(&self, tcx: TyCtxt<'_>) -> bool {
         tcx.trait_is_auto(self.def_id)
     }
-    pub(crate) fn is_label_trait(&self, tcx: TyCtxt<'_>) -> bool {
-        tcx.is_doc_label_trait(self.def_id)
-    }
     pub(crate) fn is_notable_trait(&self, tcx: TyCtxt<'_>) -> bool {
         tcx.is_doc_notable_trait(self.def_id)
     }
