+++ tests/rustdoc-html/notable-trait/notable-trait-badge-generic.rs
@@ -1,13 +1,13 @@
-#![feature(doc_label_trait)]
+#![feature(doc_notable_trait)]
 #![crate_name = "foo"]
 
-#[doc(label_trait)]
+#[doc(notable_trait)]
 pub trait Labeled {}
 
 pub trait Bound {}
 
 // A conditional impl: the badge is rendered unconditionally even though the
 // impl only holds for `T: Bound`.
-//@ has 'foo/struct.Wrapper.html' '//a[@class="impl-label-trait-full-badge"]' 'Labeled'
+//@ has 'foo/struct.Wrapper.html' '//a[@class="notable-trait-badge"]' 'Labeled'
 pub struct Wrapper<T>(pub T);
 impl<T: Bound> Labeled for Wrapper<T> {}
