--- tests/ui/feature-gates/feature-gate-doc_label_trait.rs
@@ -1,4 +1,0 @@
-#[doc(label_trait)] //~ ERROR: `#[doc(label_trait)]` is experimental
-trait SomeTrait {}
-
-fn main() {}
