--- tests/ui/feature-gates/feature-gate-doc_label_trait.stderr
@@ -1,13 +1,0 @@
-error[E0658]: `#[doc(label_trait)]` is experimental
-  --> $DIR/feature-gate-doc_label_trait.rs:1:1
-   |
-LL | #[doc(label_trait)]
-   | ^^^^^^^^^^^^^^^^^^^
-   |
-   = note: see issue #156865 <https://github.com/rust-lang/rust/issues/156865> for more information
-   = help: add `#![feature(doc_label_trait)]` to the crate attributes to enable
-   = note: this compiler was built on YYYY-MM-DD; consider upgrading it if it is out of date
-
-error: aborting due to 1 previous error
-
-For more information about this error, try `rustc --explain E0658`.
